Extracted winner reducer into another file
