+++ tictactoe/src/redux/winner/actions.js
@@ -1,0 +1,6 @@
+export const WINNER = 'WINNER';
+
+export const thereIsAWinner = squares => ({
+  type: WINNER,
+  squares
+});
