+++ tictactoe/src/redux/winner/reducer.js
@@ -1,0 +1,34 @@
+const initialState = {
+  winner: null
+};
+
+function calculateWinner(squares) {
+  const lines = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6],
+  ];
+  for (let i = 0; i < lines.length; i += 1) {
+    const [a, b, c] = lines[i];
+    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+      return squares[a];
+    }
+  }
+  return null;
+}
+
+export const thereIsAWinner = (state = initialState, action) => {
+  switch (action.type) {
+    case 'WINNER':
+      return {
+        winner: calculateWinner(action.squares)
+      };
+    default:
+      return state;
+  }
+};
